Add files via upload
## Introduction

This project is a simple implementation of a maze pathfinding algorithm in Java. The algorithm uses breadth-first search (BFS) to find the shortest path from the start point to the end point in a predefined maze. The project demonstrates basic concepts of pathfinding, data structures, and algorithms.

## Features

- Breadth-first search pathfinding
- Visualization of the pathfinding process with delays
- Simple console-based output

## Prerequisites

To run this project, you need to have the following installed:

- Java Development Kit (JDK) 8 or higher

## Setup